Reference

Tram2DKF.EARTH_GRAVITY

Tram2DKF.IDX_ACCELERATION

Tram2DKF.IDX_CURVATURE

Tram2DKF.IDX_DCURVATURE

Tram2DKF.IDX_DISTANCE

Tram2DKF.IDX_HEADING

Tram2DKF.IDX_JERK

Tram2DKF.IDX_SPEED

Tram2DKF.IDX_TIME

Tram2DKF.IDX_X_COORD

Tram2DKF.IDX_Y_COORD

Tram2DKF.Accelerate

Tram2DKF.AccelerateState

Tram2DKF.ActiveSpeedProfileSegment

Tram2DKF.ActiveTrackSegment

Tram2DKF.BacktrackingLineSearch

Tram2DKF.BacktrackingLineSearch

Tram2DKF.CartesianCA

Tram2DKF.CartesianCTRV

Tram2DKF.CartesianCV

Tram2DKF.CompositeMeasurement

Tram2DKF.CompositeMeasurement

Tram2DKF.ConstantSpeed

Tram2DKF.ConstantSpeedState

Tram2DKF.ContinuousTime

Tram2DKF.DiscreteTime

Tram2DKF.DiscretizedStateEquation

Tram2DKF.ExtendedKalmanFilter

Tram2DKF.Gaussian

Tram2DKF.Gaussian

Tram2DKF.GeneratorState

Tram2DKF.IteratedExtendedKalmanFilter

Tram2DKF.KalmanFilter

Tram2DKF.KalmanSmoother

Tram2DKF.LTIMeasurementEquation

Tram2DKF.LTIStateEquation

Tram2DKF.LinearKalmanFilter

Tram2DKF.MeasurementEquation

Tram2DKF.PolarCA

Tram2DKF.PolarCCA

Tram2DKF.PolarCTRA

Tram2DKF.PolarCTRV

Tram2DKF.PolarCV

Tram2DKF.SegmentChainer

Tram2DKF.SegmentChainer

Tram2DKF.SmoothAccelerateState

Tram2DKF.SmoothlyAccelerate

Tram2DKF.SpeedProfileSegment

Tram2DKF.SqrtGaussian

Tram2DKF.SqrtGaussian

Tram2DKF.StateEquation

Tram2DKF.Stop

Tram2DKF.StopState

Tram2DKF.StraightTrack

Tram2DKF.StraightTrackState

Tram2DKF.Time

Tram2DKF.TrackCurvature

Tram2DKF.TrackSegment

Tram2DKF.TrackTurn

Tram2DKF.TrackTurnState

Tram2DKF.TrajectoryDrive

Tram2DKF.TramMotionModel

Tram2DKF.UncertainValue

Tram2DKF.AlwaysFeasible

Tram2DKF.Euler

Tram2DKF.IdentityStepping

Tram2DKF.RK4

Tram2DKF.activate

Tram2DKF.activate

Tram2DKF.activate

Tram2DKF.activate

Tram2DKF.activate

Tram2DKF.activate

Tram2DKF.backward_step

Tram2DKF.covariance

Tram2DKF.covariance

Tram2DKF.curvature

Tram2DKF.curvature

Tram2DKF.data_step

Tram2DKF.data_step

Tram2DKF.data_step

Tram2DKF.direct_forward_step

Tram2DKF.direct_forward_step

Tram2DKF.discretize

Tram2DKF.dlerp

Tram2DKF.drive

Tram2DKF.drive

Tram2DKF.drive

Tram2DKF.drive

Tram2DKF.forward_step

Tram2DKF.forward_step

Tram2DKF.forward_step

Tram2DKF.iekf_finalize

Tram2DKF.iekf_finalize

Tram2DKF.iekf_gauss_newton_step!

Tram2DKF.inner_generator_loop!

Tram2DKF.innovation_data_step

Tram2DKF.innovation_data_step

Tram2DKF.is_working

Tram2DKF.is_working

Tram2DKF.kick!

Tram2DKF.lerp

Tram2DKF.linearize

Tram2DKF.linearize

Tram2DKF.logpdf

Tram2DKF.logpdf

Tram2DKF.mean

Tram2DKF.mean

Tram2DKF.pdf

Tram2DKF.pdf

Tram2DKF.render_trip

Tram2DKF.simulate_accelerometer

Tram2DKF.simulate_gnss

Tram2DKF.simulate_gyro

Tram2DKF.simulate_odometry

Tram2DKF.EARTH_GRAVITY — Constant

Earth gravitational acceleration, in m/s^2.

Tram2DKF.IDX_ACCELERATION — Constant

Index into TramMotionModel state: Forward acceleration of the tram, in m/s^2

Tram2DKF.IDX_CURVATURE — Constant

Index into TramMotionModel state: Track curvature, in 1/m

Tram2DKF.IDX_DCURVATURE — Constant

Index into TramMotionModel state: Track curvature derivative, in 1/m^2

Tram2DKF.IDX_DISTANCE — Constant

Index into TramMotionModel state: Distance travelled since start

Tram2DKF.IDX_HEADING — Constant

Index into TramMotionModel state: Track heading, in radians

Tram2DKF.IDX_JERK — Constant

Index into TramMotionModel state: Forward jerk of the tram, in m/s^3

Tram2DKF.IDX_SPEED — Constant

Index into TramMotionModel state: Forward speed of the tram, in m/s

Tram2DKF.IDX_TIME — Constant

Index into TramMotionModel state: Time elapsed since start, in seconds

Tram2DKF.IDX_X_COORD — Constant

Index into TramMotionModel state: X coordinate of the tram, in metres

Tram2DKF.IDX_Y_COORD — Constant

Index into TramMotionModel state: Y coordinate of the tram, in metres

Tram2DKF.Accelerate — Type

Accelerate{NumT}

The tram should accelerate from its current speed to a new to_speed at a given acceleration.

Tram2DKF.AccelerateState — Type

AccelerateState{NumT}

Generate linearly ramped speed profile:

at from_time the speed is from_speed

at to_time the speed is to_speed

acceleration is provided for convenience; it has the correct sign.

Tram2DKF.ActiveSpeedProfileSegment — Type

ActiveSpeedProfileSegment

Generates speed profile information for its specific segment.

Interface

drive(segment::ActiveSpeedProfileSegment, time, position, speed, acceleration)::Union{TrajectoryDrive, Nothing}

Returns the speed and acceleration at the given point on the speed profile. The position and time values are not relative wrt. the segment start, they are relative to the start of the entire journey.

The method needs to return nothing when this segment has ended.

distance(segment::ActiveSpeedProfileSegment)

Get the distance travelled during this segment in metres.

duration(segment::ActiveSpeedProfileSegment)

Get the duration in seconds that it will take the tram to traverse this segment.

end_state(segment::ActiveSpeedProfileSegment)::TrajectoryDrive

Return the tram command at the segment end.

Tram2DKF.ActiveTrackSegment — Type

ActiveTrackSegment

Generates curvature information for a specific segment of a track.

Interface

curvature(segment::ActiveTrackSegment, position)::Union{TrackCurvature, Nothing}

Returns the local curvature at position metres from track origin. The position value is not relative wrt. the segment start.

The method needs to return nothing when position is after the end of the track segment. However, it should return zero position when the position is before the start of the segment.

distance(segment::ActiveTrackSegment)::NumT

Returns the length of this segment in metres.

end_state(segment::ActiveTrackSegment)::TrackCurvature

Return the track geometry at the segment end.

Tram2DKF.BacktrackingLineSearch — Type

BacktrackingLineSearch{T <: Real}

Optimize step length by iteratively shortening the step and checking if the Armijo condition holds.

The Armijo condition holds if the function has decreased enough compared to its derivative at the origin point.

Additionally, the function can check if the resulting vector falls into a feasible region. The check is provided by a user-defined function.

Tram2DKF.BacktrackingLineSearch — Method

(bls::BacktrackingLineSearch)(V, x0, xstep0)

Optimize step length via backtracking line search.

The function will find such a step to satisfy the Armijo condition when minimizing function V in the direction xstep0 when coming from point x0.

This function assumes that the x0is feasible (thebls.isfeasiblefunction shall return true atx0`).

The new step vector is returned.

Tram2DKF.CartesianCA — Type

CartesianCA # cartesian, constant acceleration

Represents an accelerating point mass. Velocity and acceleration are parameterized in cartesian coordinates.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... X speed

x[4] ... Y speed

x[5] ... X acceleration

x[6] ... Y acceleration

Tram2DKF.CartesianCTRV — Type

CartesianCTRV # cartesian, constant turn rate, constant velocity

Represents an rotating rigid body moving at a constant speed magnitude. Velocity is parameterized in cartesian coordinates and is continuously rotating.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... X speed

x[4] ... Y speed

x[5] ... yaw rate

Tram2DKF.CartesianCV — Type

CartesianCV # cartesian, constant velocity

Represents an point mass moving at a constant speed. Velocity is parameterized in cartesian coordinates.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... X speed

x[4] ... Y speed

Tram2DKF.CompositeMeasurement — Type

CompositeMeasurement

Glues multiple independent measurements models into one measurement vector.

Tram2DKF.CompositeMeasurement — Method

(msr::CompositeMeasurement)(x, u)

Evaluate specified sub-measurements and concatenate their results into one large measurement vector.

Tram2DKF.ConstantSpeed — Type

ConstantSpeed{NumT}

The tram should move at a constant speed until it travels a given distance.

Tram2DKF.ConstantSpeedState — Type

ConstantSpeedState{NumT}

Generate a constant speed profile until the to_point time mark.

Tram2DKF.ContinuousTime — Type

ContinuousTime

Marker indicating that the StateEquation operates in continuous time.

Tram2DKF.DiscreteTime — Type

DiscreteTime

Marker indicating that the StateEquation operates in discrete time.

Tram2DKF.DiscretizedStateEquation — Type

DiscretizedStateEquation{SubmodelT <: StateEquation{ContinuousTime}, DiscretizationMethodT, TimeT}

Helper object that wraps a continuous-time system and behaves like a discrete-time system.

Tram2DKF.ExtendedKalmanFilter — Type

Extended Kalman filter

The EKF allows you to estimate the states of a nonlinear dynamical system under white Gaussian noise.

EKF deals with the nonlinearities by linearizing the system at the current state estimate and then effectively applying the linear Kalman filter on that.

Tram2DKF.Gaussian — Type

Gaussian{MeanT,CovT} <: UncertainValue

Represents a multidimensional normally distributed random variable.

Tram2DKF.Gaussian — Method

Gaussian(g::UncertainValue)

Create a moment-matched Gaussian random variable for the given random variable g.

Tram2DKF.GeneratorState — Type

GeneratorState{NumT, SystemT, TimeT}

Helper structure for carrying track generator state between functions.

Tram2DKF.IteratedExtendedKalmanFilter — Type

IteratedExtendedKalmanFilter{StepSizeControlT, StepNormT <: Real}

The IEKF allows you to estimate the states of a nonlinear dynamical system under white Gaussian noise.

Compared to the classical EKF, the IEKF may work better for systems with strongly nonlinear measurement functions. This is because the IEKF actually finds the maximum aposteriori estimate when combining the state estimate with measurements.

The IEKF was implemented based on the article https://ieeexplore.ieee.org/document/10705771 . This article also illustrates that the IEKF is equivalent to a Gauss-Newton iteration that maximizes some criterion. The GN method is then extended to include step-size control to speed up convergence.

Tram2DKF.KalmanFilter — Type

KalmanFilter

Algorithm for fusing measurements from multiple sensors and for estimating a state of a system whose state is measured.

Interface

forward_step(kf::KalmanFilter,
             model::StateEquation{DiscreteTime},
             prev_state::UncertainValue,
             input,
             process_noise::UncertainValue)::UncertainValue

Estimate the state of model at the next time step.

The system is assumed to be in state prev_state with the specified input applied over the entire sampling period. Additionally, there is an unknown random disturbance described by process_noise.

Return the Gaussian belief about the state at the next time step.

This represents the classic Kalman "time step" / "forward step".

data_step(kf::KalmanFilter,
          model::StateEquation{DiscreteTime},
          prior::UncertainValue,
          input,
          observation::UncertainValue)::UncertainValue

Correct the current model state estimate prior using the observation.

The system is assumed to be in state prior with the specified input applied over the current sampling period.

Return the updated Gaussian belief about the state at the current time step.

This represents the classic Kalman "data step".

Tram2DKF.KalmanSmoother — Type

KalmanSmoother

Extension of KalmanFilter that supports smoothing (propagation of information back in time).

Interface

backward_step(kf::KalmanFilter,
              model::StateEquation{DiscreteTime},
              current_posterior::UncertainValue,
              next_prior::UncertainValue,
              next_smoothed::UncertainValue)::UncertainValue

Smooth the state of model at the current time step.

This represents the classic Rauch-Tung-Striebel smoothing algorithm that allows you to propagate information back in time. Given:

the estimate of state at time k given data up to time k (current_posterior),

the estimate of state at time k+1 given data up to time k (next_prior),

the estimate of state at time k+1 given data up to time l > k (next_smoothed),

the algorithm gives you an estimate at time k based on data up to time l > k.

Tram2DKF.LTIMeasurementEquation — Type

LTIMeasurementEquation

Linear time-invariant measurement model.

Tram2DKF.LTIStateEquation — Type

LTIStateEquation{T <: Time}

Linear time-invariant state-space model.

Tram2DKF.LinearKalmanFilter — Type

LinearKalmanFilter

The LKF allows you to optimally estimate states of linear dynamical systems under white Gaussian noise.

Tram2DKF.MeasurementEquation — Type

MeasurementEquation

Model of a system state measurement.

The model is represented as an explicit function. Implicit measurement equations are not supported.

Interface

Users can expect values of this type to have the following methods:

y = (eq::MeasurementEquation)(x, u)

Predict the measurement value based on the current state and return it as a vector.

ninputs(eq::StateEquation)

Expected length of the control vector u.

nstates(eq::StateEquation)

Expected length of the state vector x.

noutputs(eq::StateEquation)

Expected length of the measurement vector y.

Tram2DKF.PolarCA — Type

PolarCA # polar, constant acceleration

Represents an accelerating point mass. Velocity is parameterized in polar coordinates; acceleration is assumed to only act in the forward direction.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... forward velocity

x[4] ... forward acceleration

x[5] ... heading

Tram2DKF.PolarCCA — Type

PolarCCA # polar, constant curvature, constant acceleration

Represents an accelerating rigid body moving along a circular curve. Velocity is parameterized in polar coordinates; acceleration is assumed to only act in the forward direction.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... forward velocity

x[4] ... forward acceleration

x[5] ... heading

x[6] ... track curvature

Tram2DKF.PolarCTRA — Type

PolarCTRA # polar, constant turn rate, constant acceleration

Represents an accelerating 2D rigid body moving along a clothoidal path. Velocity is parameterized in polar coordinates; acceleration is assumed to only act in the forward direction.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... forward velocity

x[4] ... forward acceleration

x[5] ... heading

x[6] ... yaw rate

Tram2DKF.PolarCTRV — Type

PolarCTRV # polar, constant turn rate, constant velocity

Represents a 2D rigid body moving along a circular path along a constant speed. Velocity is parameterized in polar coordinates.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... forward velocity

x[4] ... heading

x[5] ... yaw rate

Tram2DKF.PolarCV — Type

PolarCV # polar, constant velocity

Represents a point mass moving at a constant velocity. Velocity is parameterized in polar coordinates.

State vector layout

x[1] ... X position

x[2] ... Y position

x[3] ... forward velocity

x[4] ... heading

Tram2DKF.SegmentChainer — Type

SegmentChainer{SegmentCollection}

Helper object for chaining track and speed profile segments.

Tram2DKF.SegmentChainer — Method

SegmentChainer(segments, arg_pack)

Create an object for chaining of segments of differing types.

Arguments

The segments argument should be an iterable of segment descriptions (e.g. TrackSegments or SpeedProfileSegments)

The arg_pack argument should be a tuple of additional state arguments to pass to the activate function

Tram2DKF.SmoothAccelerateState — Type

SmoothAccelerateState{NumT}

Generate ramped speed profile with limited jerk.

Tram2DKF.SmoothlyAccelerate — Type

SmoothlyAccelerate{NumT}

The tram should accelerate from its current speed to a new to_speed at a given acceleration, while the rate of acceleration change is capped by a given jerk.

Tram2DKF.SpeedProfileSegment — Type

SpeedProfileSegment

Specifies a part of a tram speed profile.

Interface

activate(segment::SpeedProfileSegment, time, position, speed, acceleration)::ActiveSpeedProfileSegment

Returns a description of the segment starting from a given tram motion state.

Tram2DKF.SqrtGaussian — Type

SqrtGaussian

Represents a multidimensional normally distributed random variable.

Compared to Gaussian, this struct represents the covariance in a LLᵀ factorized form. Some Kalman filter implementations can exploit this to be more numerically stable.

Tram2DKF.SqrtGaussian — Method

Gaussian(g::UncertainValue)

Create a moment-matched Gaussian random variable for the given random variable g.

Tram2DKF.StateEquation — Type

StateEquation{T <: Time}

State-space dynamics model for an abstract system.

The model is represented by a first-order explicit vector ODE. DAE models are not supported.

Whether the system is evolving in continuous or discrete time is determined by the T type argument. See DiscreteTime and ContinuousTime.

Interface

Users can expect values of this type to have the following methods:

(eq::StateEquation)(x, u)

Advance the state of the system:

continuous-time systems return the derivative of the state at the current time

discrete-time systems return the new state at the next time instant

Dependence on time is not explicitly modelled. If needed, the time can be obtained by introducing an additional state with derivative equal to 1.

ninputs(eq::StateEquation)

Expected length of the control vector u.

nstates(eq::StateEquation)

Expected length of the state vector x.

Tram2DKF.Stop — Type

Stop{NumT}

The tram should stand still for duration seconds.

Tram2DKF.StopState — Type

StopState{NumT}

Generate zero speed until the end_at time mark.

Tram2DKF.StraightTrack — Type

StraightTrack{NumT}

Describes a distance metres long segment of a straight track (i.e. zero curvature).

Tram2DKF.StraightTrackState — Type

StraightTrackState{NumT}

Generates zero curvature between from_point and to_point track positions.

Tram2DKF.Time — Type

Time

Marker type for discriminating between continuous-time and discrete-time models. Typically one of DiscreteTime and ContinuousTime.

See also StateEquation.

Tram2DKF.TrackCurvature — Type

TrackCurvature{NumT}

Represents the local geometry of a tram track.

Tram2DKF.TrackSegment — Type

TrackSegment

Specifies a part of a tram track geometry.

Interface

activate(segment::TrackSegment, position)::ActiveTrackSegment

Returns a description of the segment given its starting position in metres.

Tram2DKF.TrackTurn — Type

TrackTurn{NumT <: Real}

Describes a track curve, optionally including clothoid transition segments.

Tram2DKF.TrackTurnState — Type

TrackTurnState

Precomputed trapezoidal curvature profile for a track turn.

max_curvature is the maximum curvature encountered.

between transition_in_start and arc_start the curvature will be linearly increased

between arc_start and transition_out_start the curvature will be constant (max_curvature)

between transition_out_start and turn_end the curvature will be linearly decreased

Tram2DKF.TrajectoryDrive — Type

TrajectoryDrive{NumT}

Represents the local shape of the speed profile.

Tram2DKF.TramMotionModel — Type

TramMotionModel

State space model used for generating the tram trajectory.

The model basically a combination of three parts:

X/Y coordinates of the tram

longitudinal dynamics: distance, speed, acceleration

track geometry: heading, curvature and its derivative

The state trajectories of this model should contain at least straight lines, circular arcs and clothoids.

State vector layout

x[IDX_TIME] ... Time elapsed since start, in seconds

x[IDX_DISTANCE] ... Distance travelled since start

x[IDX_X_COORD] ... X coordinate of the tram, in metres

x[IDX_Y_COORD] ... Y coordinate of the tram, in metres

x[IDX_SPEED] ... Forward speed of the tram, in m/s

x[IDX_ACCELERATION] ... Forward acceleration of the tram, in m/s^2

x[IDX_HEADING] ... Track heading, in radians

x[IDX_CURVATURE] ... Track curvature, in 1/m

x[IDX_DCURVATURE] ... Track curvature derivative, in 1/m^2

Tram2DKF.UncertainValue — Type

UncertainValue

Represents a randomly distributed variable with a well-defined mean and covariance.

Interface

mean(variable::UncertainValue)::Vector{<:Real}

Return the expected value of the random variable.

covariance(variable::UncertainValue)::Matrix{<:Real}

Return the covariance (second central moment) of the random variable.

pdf(variable::UncertainValue, value::Vector{<:Real})::Real

Evaluate the probability density function at the given point.

logpdf(variable::UncertainValue, value::Vector{<:Real})::Real

Evaluate the logarithm of the probability density function at the given point. This function can be more numerically stable than pdf() when small probabilities are involved (e.g. tails of a Gaussian).

Tram2DKF.AlwaysFeasible — Method

AlwaysFeasible(x)

Noop feasibility check - all X are feasible.

Tram2DKF.Euler — Method

Euler(f,x,u,dt)

Run a single step of the Euler discretization method.

Return the state at which the system specified by the ODE f will be after dt seconds if the current state is x and control is u.

Tram2DKF.IdentityStepping — Method

IdentityStepping(V, x0, xstep0)

Noop step size control: pass through the initial step size xstep0.

Tram2DKF.RK4 — Method

RK4(f,x,u,dt)

Run a single step of the Runge-Kutta order 4 discretization method.

Return the state at which the system specified by the ODE f will be after dt seconds if the current state is x and control is u.

Tram2DKF.activate — Method

activate(acc::Accelerate, time::NumT, pos, speed, accel) where {NumT <: AbstractFloat}

Return a description of the segment starting from a given tram motion state.

Tram2DKF.activate — Method

activate(spd::ConstantSpeed, time::NumT, pos, speed, accel) where {NumT <: AbstractFloat}

Return a description of the segment starting from a given tram motion state.

Tram2DKF.activate — Method

activate(acc::SmoothlyAccelerate, time::NumT, pos, speed, accel) where {NumT <: AbstractFloat}

Return a description of the segment starting from a given tram motion state.

Tram2DKF.activate — Method

activate(stop::Stop, time::NumT, pos, speed, accel) where {NumT <: AbstractFloat}

Return a description of the segment starting from a given tram motion state.

Tram2DKF.activate — Method

activate(seg::StraightTrack, pos::NumT) where {NumT <: AbstractFloat}

Returns a description of the segment given its starting position in metres.

Tram2DKF.activate — Method

activate(seg::TrackTurn, pos::NumT) where {NumT <: AbstractFloat}

Returns a description of the segment given its starting position in metres.

Tram2DKF.backward_step — Method

backward_step(::LinearKalmanFilter,
              model::LTIStateEquation{DiscreteTime},
              current_posterior::UncertainValue,
              next_prior::UncertainValue,
              next_smoothed::UncertainValue)

Perform a smoothing step of a linear Rauch-Tung-Striebel smoother.

This allows you to propagate information from new measurements back in time and make the estimates of past states less noisy.

Tram2DKF.covariance — Method

covariance(g::Gaussian)

Return the covariance of the Gaussian.

Tram2DKF.covariance — Method

covariance(g::SqrtGaussian)

Return the covariance of the Gaussian.

Tram2DKF.curvature — Method

curvature(seg::TrackTurnState{NumT}, pos) where {NumT}

Return the local curvature at position metres from track origin. The position value is not relative wrt. the segment start.

However, return nothing when position is after the end of the track segment.

Tram2DKF.curvature — Method

curvature(seg::TrackTurnState{NumT}, pos) where {NumT}

Return the local curvature at position metres from track origin. The position value is not relative wrt. the segment start.

However, return nothing when position is after the end of the track segment.

Tram2DKF.data_step — Method

data_step(::ExtendedKalmanFilter,
          g::MeasurementEquation,
          prior::UncertainValue,
          input,
          observation::UncertainValue)

Perform a data step of an extended Kalman filter.

This allows you to reduce the undertainty of a state estimate by incorporating new information from a state measurement.

Tram2DKF.data_step — Method

data_step(iekf::IteratedExtendedKalmanFilter,
          g::MeasurementEquation,
          prior::UncertainValue,
          input,
          observation::UncertainValue)

Perform a data step of an iterated extended Kalman filter.

This allows you to reduce the undertainty of a state estimate by incorporating new information from a state measurement.

This method actually computes the maximum aposteriori estimate of the state. Uncertainty propagation is handled by linearizing the system around this MAP estimate and then propagating the Gaussians just like the linear KF does.

Tram2DKF.data_step — Method

data_step(::LinearKalmanFilter,
          model::LTIMeasurementEquation,
          prior::UncertainValue,
          input,
          observation::UncertainValue)

Perform a data step of a linear Kalman filter.

This allows you to reduce the undertainty of a state estimate by incorporating new information from a linear state measurement.

Tram2DKF.direct_forward_step — Method

direct_forward_step(
    model::LTIStateEquation{DiscreteTime},
    prev_state::SqrtGaussian,
    new_state,
    process_noise::SqrtGaussian)

Compute the covariance of a state estimate at time k+1 given data up to time k. Then attach it to a precomputed mean new_state and return it as a new belief.

Compared to the direct_forward_step above, this method uses the QR factorization to compute the final covariance. This may be more numerically stable.

For details about the math, see T. M. Chin, „Square-root formulas for Kalman filter, information filter, and RTS smoother: Links via boomerang prediction residual".

Tram2DKF.direct_forward_step — Method

direct_forward_step(
    model::LTIStateEquation{DiscreteTime},
    prev_state::UncertainValue,
    new_state,
    process_noise::UncertainValue)

Compute the covariance of a state estimate at time k+1 given data up to time k. Then attach it to a precomputed mean new_state and return it as a new belief.

Tram2DKF.discretize — Function

discretize(model::StateEquation{ContinuousTime}, method, sampling_time, subsamples::Int = 1)

Discretize the given model using the discretization method with given sampling_time.

Additionally, each sampling_time interval can be internally split into subsamples sub-intervals over which will the discretization method` be repeatedly invoked.

Tram2DKF.dlerp — Method

dlerp(x0, y0, x1, y1)

Calculate the slope of a line going through points (x0, y0) and (x1, y1).

This function is the derivative of dlerp with respect to the xq argument.

See also dlerp.

Tram2DKF.drive — Method

drive(acc::AccelerateState{NumT}, time, pos, speed, accel) where {NumT}

Return the speed and acceleration at the given point on the speed profile. The position and time values are not relative wrt. the segment start, they are relative to the start of the entire journey.

Return nothing when this segment has ended.

Tram2DKF.drive — Method

drive(spd::ConstantSpeedState{NumT}, time, pos, speed, accel) where {NumT}

Return the speed and acceleration at the given point on the speed profile. The position and time values are not relative wrt. the segment start, they are relative to the start of the entire journey.

Return nothing when this segment has ended.

Tram2DKF.drive — Method

drive(acc::SmoothAccelerateState{NumT}, time, pos, speed, accel) where {NumT}

Return the speed and acceleration at the given point on the speed profile. The position and time values are not relative wrt. the segment start, they are relative to the start of the entire journey.

Return nothing when this segment has ended.

Tram2DKF.drive — Method

drive(stop::StopState{NumT}, time, pos, speed, accel) where {NumT}

Return the speed and acceleration at the given point on the speed profile. The position and time values are not relative wrt. the segment start, they are relative to the start of the entire journey.

Return nothing when this segment has ended.

Tram2DKF.forward_step — Method

forward_step(::ExtendedKalmanFilter,
             f::StateEquation{DiscreteTime},
             prev_state::UncertainValue,
             input,
             process_noise::UncertainValue)

Perform a time step of an extended Kalman filter.

This will estimate the state of model at the next time step given the belief of prev_state and knowledge of input.

Tram2DKF.forward_step — Method

forward_step(::IteratedExtendedKalmanFilter,
             f::StateEquation{DiscreteTime},
             prev_state::UncertainValue,
             input,
             process_noise::UncertainValue)

Perform a time step of an iterated extended Kalman filter.

This will estimate the state of model at the next time step given the belief of prev_state and knowledge of input.

Given that the IEKF does not iterate the time update, this is fully equivalent to the EKF forward step.

Tram2DKF.forward_step — Method

forward_step(::LinearKalmanFilter,
             model::LTIStateEquation{DiscreteTime},
             prev_state::UncertainValue,
             input,
             process_noise::UncertainValue)

Perform a time step of a linear Kalman filter.

This will estimate the state of model at the next time step given the belief of prev_state and knowledge of input.

Tram2DKF.iekf_finalize — Method

iekf_finalize(
    ad_cache,
    g::MeasurementEquation,
    prior::SqrtGaussian,
    map_estimate,
    input,
    observation::SqrtGaussian)

Compute the final state estimate covariance around the MAP estimate.

This variant uses square-root-filtering algoritms for increased accuracy.

Tram2DKF.iekf_finalize — Method

iekf_finalize(
    ad_cache,
    g::MeasurementEquation,
    prior::UncertainValue,
    map_estimate,
    input,
    observation::UncertainValue)

Compute the final state estimate covariance around the MAP estimate.

Tram2DKF.iekf_gauss_newton_step! — Method

iekf_gauss_newton_step!(
    ad_cache,
    g::MeasurementEquation,
    prior::UncertainValue,
    current_est,
    input,
    observation::UncertainValue)

Perform one iteration of the internal Gauss-Newton-like method.

This method does not perform step size control. Instead, the unscaled step is returned.

See https://ieeexplore.ieee.org/document/10705771 for details.

Tram2DKF.inner_generator_loop! — Method

inner_generator_loop!(gen::GeneratorState, track::ActiveTrackSegment, speed_profile::ActiveSpeedProfileSegment)

Hot loop of the trajectory generator.

This function is factored out in order to have type-stable track and speed_profile arguments. The Julia compiler can then avoid calling into dynamic dispatch.

The function will return whenever the track or speed_profile generators start returning nothing. The calling function should then fetch next segments where needed and then call this function again.

Tram2DKF.innovation_data_step — Method

innovation_data_step(
    model::LTIMeasurementEquation,
    prior::SqrtGaussian,
    innovation,
    observation::SqrtGaussian)

Compute a corrected state estimate given a new measurement.

This method is intended as a reusable piece of data_step. It takes in a innovation, which is the difference between the actual measured value and the value predicted by the state estimate prior. Thanks to this, it can be, in principle, applied to nonlinear models as well.

Compared to the innovation_data_step above, this method uses the QR factorization to compute the final covariance. This may be more numerically stable.

For details about the math, see T. M. Chin, „Square-root formulas for Kalman filter, information filter, and RTS smoother: Links via boomerang prediction residual".

Tram2DKF.innovation_data_step — Method

innovation_data_step(
    model::LTIMeasurementEquation,
    prior::UncertainValue,
    innovation,
    observation::UncertainValue)

Compute a corrected state estimate given a new measurement.

This method is intended as a reusable piece of data_step. It takes in a innovation, which is the difference between the actual measured value and the value predicted by the state estimate prior. Thanks to this, it can be, in principle, applied to nonlinear models as well.

Tram2DKF.is_working — Method

is_working(seg::ActiveSpeedProfileSegment, arg_pack::NTuple{4, <:AbstractFloat})

Check that the current speed profile segment has not ended.

Tram2DKF.is_working — Method

is_working(seg::ActiveTrackSegment, arg_pack::NTuple{1, <:AbstractFloat})

Check that the current track segment has not ended.

Tram2DKF.kick! — Method

kick!(chainer::SegmentChainer, arg_pack)

Ensure that the active segment generator is still working. If not, obtain a generator for the next segment. If there no more segments, return false. On the other hand, if the generator in the active_segment field was successfully refreshed, return true.

Tram2DKF.lerp — Method

lerp(x0, y0, x1, y1, xq)

Perform linear interpolation between two points.

The interpolation line is specified using the points (x0, y0) and (x1, y1). The function then returns the y-coordinate of the point with x-coordinate equal to xq.

See also dlerp.

Tram2DKF.linearize — Method

linearize(g::MeasurementEquation, x, u)

Linearize the given nonlinear measurement model g at the operating point specified by state x and control u.

Tram2DKF.linearize — Method

linearize(f::StateEquation{TimeT}, x, u) where {TimeT <: Time}

Linearize the given nonlinear dynamic model f at the operating point specified by state x and control u.

Tram2DKF.logpdf — Method

logpdf(g::Gaussian, value)

Evaluate the logarithm of the probability density function at the given point.

Tram2DKF.logpdf — Method

logpdf(g::SqrtGaussian, value)

Evaluate the logarithm of the probability density function at the given point.

Tram2DKF.mean — Method

mean(g::Gaussian)

Return the mean of the Gaussian.

Tram2DKF.mean — Method

mean(g::SqrtGaussian)

Return the mean of the Gaussian.

Tram2DKF.pdf — Method

pdf(g::Gaussian, value)

Evaluate the probability density function at the given point.

Tram2DKF.pdf — Method

pdf(g::SqrtGaussian, value)

Evaluate the probability density function at the given point.

Tram2DKF.render_trip — Function

render_trip(track_segments::AbstractVector{<:TrackSegment}, speed_segments::AbstractVector{<:SpeedProfileSegment}, dt::Float64, subsamples::Int = 1)

Simulate a state-space trajectory of a tram going along track_segments and following a speed profile composed of speed_segments.

The simulation will return one sample for every dt seconds elapsed. However, the ODE can be invoked more frequently that that in order to increase accuracy. To do so, provide a subsamples value greater than 1.

The state trajectory is returned as a vector of (state) vectors.

Tram2DKF.simulate_accelerometer — Method

simulate_accelerometer(states, wn_std)

Simulate 3-axis accelerometer measurements for the trajectory defined by the states sequence.

The accelerometer is assumed to be aligned with the vehicle body (X forward, Y left, Z up). Additionally, the measurements are corrupted by a white Gaussian noise with standard deviation equal to wn_std.

Tram2DKF.simulate_gnss — Method

simulate_gnss(states, altitude, cov_matrix)

Simulate GNSS measurements for the trajectory defined by the states sequence.

The GNSS will return directly the X and Y state components in metres. Additionally, it will return the specified altitude. Finally, all the components are corrupted by a white Gaussian noise with cov_matrix covariance matrix.

Tram2DKF.simulate_gyro — Method

simulate_gyro(states, wn_std)

Simulate 3-axis gyro measurements for the trajectory defined by the states sequence.

The gyro is assumed to be aligned with the vehicle body (X forward, Y left, Z up). Additionally, the measurements are corrupted by a white Gaussian noise with standard deviation equal to wn_std.

Tram2DKF.simulate_odometry — Method

simulate_odometry(states, granularity)

Simulate odometry measurements for the trajectory defined by the states sequence.

The odometer is assumed to be counting teeth on a wheel encoder. Thus, a position quantization is simulated. However, no slip is simulated.

The returned distances are still in metres.